+++ sample-code/src/openai.ts
